Sped up the object deletion function
It now uses the width of the sprite to calculate when an object has gone off the screen and works for both the ground and the trees themselves. This lets us spawn objects MUCH closer to the player. May need some tweaking when it comes to multiple objects later on*